pyApp/requirements.txt
blinker==1.6.2
click==8.1.3
enum-compat==0.0.3
filelock==3.12.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JSON==0.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
packaging==23.1
Pillow==9.5.0
psutil==5.9.5
six==1.16.0
sympy==1.11.1
torch==2.0.0
torch-model-archiver==0.7.1
torchserve==0.7.1
typing_extensions==4.5.0
Werkzeug==2.3.3
